Add authentification for administrator pages.

+++ routes/web.php
@@ -22,13 +22,13 @@ Route::get('/', function () {
  |--------------------------------------------------------------------------
  */
 
-Route::get('/equipements', 'EquipmentController@index')->name('equipements.index');
-Route::post('/equipements', 'EquipmentController@store')->name('equipements.store');
-Route::get('/equipements/creer', 'EquipmentController@create')->name('equipements.create');
-Route::get('/equipements/{equipement}', 'EquipmentController@show')->name('equipements.show');
-Route::get('/equipements/{equipement}/modifier', 'EquipmentController@edit')->name('equipements.edit');
-Route::patch('/equipements/{equipement}/', 'EquipmentController@update')->name('equipements.update');
-Route::delete('/equipements/{equipement}', 'EquipmentController@destroy')->name('equipements.destroy');
+Route::middleware('auth:web')->get('/equipements', 'EquipmentController@index')->name('equipements.index');
+Route::middleware('auth:web')->post('/equipements', 'EquipmentController@store')->name('equipements.store');
+Route::middleware('auth:web')->get('/equipements/creer', 'EquipmentController@create')->name('equipements.create');
+Route::middleware('auth:web')->get('/equipements/{equipement}', 'EquipmentController@show')->name('equipements.show');
+Route::middleware('auth:web')->get('/equipements/{equipement}/modifier', 'EquipmentController@edit')->name('equipements.edit');
+Route::middleware('auth:web')->patch('/equipements/{equipement}/', 'EquipmentController@update')->name('equipements.update');
+Route::middleware('auth:web')->delete('/equipements/{equipement}', 'EquipmentController@destroy')->name('equipements.destroy');
 
 // Routes for authentification (register page is disabled)
 Auth::routes(['register' => false]);
@@ -48,21 +48,21 @@ Route::resource('reserver', 'ReserverController');
  |--------------------------------------------------------------------------
  */
 
-Route::get('/categories', 'CategorieController@index')->name('categorie.index');
-Route::post('/categories', 'CategorieController@store')->name('categorie.store');
-Route::get('/categories/creer', 'CategorieController@create')->name('categorie.create');
-Route::get('/categories/{categorie}/modifier', 'CategorieController@edit')->name('categorie.edit');
-Route::patch('/categories/{categorie}/', 'CategorieController@update')->name('categorie.update');
+Route::middleware('auth:web')->get('/categories', 'CategorieController@index')->name('categorie.index');
+Route::middleware('auth:web')->post('/categories', 'CategorieController@store')->name('categorie.store');
+Route::middleware('auth:web')->get('/categories/creer', 'CategorieController@create')->name('categorie.create');
+Route::middleware('auth:web')->get('/categories/{categorie}/modifier', 'CategorieController@edit')->name('categorie.edit');
+Route::middleware('auth:web')->patch('/categories/{categorie}/', 'CategorieController@update')->name('categorie.update');
 
 /*
  |--------------------------------------------------------------------------
  | Utilisateurs Routes
  |--------------------------------------------------------------------------
  */
-Route::get('/utilisateurs', 'UserController@index')->name('utilisateurs.index');
-Route::post('/utilisateurs', 'UserController@store')->name('utilisateurs.store');
-Route::get('/utilisateurs/creer', 'UserController@create')->name('utilisateurs.create');
-Route::get('/utilisateurs/{utilisateur}', 'UserController@show')->name('utilisateurs.show');
-Route::get('/utilisateurs/{utilisateur}/modifier', 'UserController@edit')->name('utilisateurs.edit');
-Route::patch('/utilisateurs/{utilisateur}/', 'UserController@update')->name('utilisateurs.update');
-Route::delete('/utilisateurs/{utilisateur}', 'UserController@destroy')->name('utilisateurs.destroy');
+Route::middleware('auth:web')->get('/utilisateurs', 'UserController@index')->name('utilisateurs.index');
+Route::middleware('auth:web')->post('/utilisateurs', 'UserController@store')->name('utilisateurs.store');
+Route::middleware('auth:web')->get('/utilisateurs/creer', 'UserController@create')->name('utilisateurs.create');
+Route::middleware('auth:web')->get('/utilisateurs/{utilisateur}', 'UserController@show')->name('utilisateurs.show');
+Route::middleware('auth:web')->get('/utilisateurs/{utilisateur}/modifier', 'UserController@edit')->name('utilisateurs.edit');
+Route::middleware('auth:web')->patch('/utilisateurs/{utilisateur}/', 'UserController@update')->name('utilisateurs.update');
+Route::middleware('auth:web')->delete('/utilisateurs/{utilisateur}', 'UserController@destroy')->name('utilisateurs.destroy');
